upgrading modules to latest pre-release (wip)

+++ source/dotnet-cqrs-petclinic-rest/PetClinic.Application/Owners/DeleteOwner/DeleteOwnerCommandHandler.cs
@@ -3,10 +3,11 @@ using System.Threading;
 using System.Threading.Tasks;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
+using PetClinic.Domain.Common.Exceptions;
 using PetClinic.Domain.Repositories;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
-[assembly: IntentTemplate("Intent.Application.MediatR.CommandHandler", Version = "1.0")]
+[assembly: IntentTemplate("Intent.Application.MediatR.CommandHandler", Version = "2.0")]
 
 namespace PetClinic.Application.Owners.DeleteOwner
 {
@@ -22,11 +23,16 @@ namespace PetClinic.Application.Owners.DeleteOwner
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
-        public async Task<Unit> Handle(DeleteOwnerCommand request, CancellationToken cancellationToken)
+        public async Task Handle(DeleteOwnerCommand request, CancellationToken cancellationToken)
         {
             var existingOwner = await _ownerRepository.FindByIdAsync(request.Id, cancellationToken);
+            if (existingOwner is null)
+            {
+                throw new NotFoundException($"Could not find Owner '{request.Id}'");
+            }
+
             _ownerRepository.Remove(existingOwner);
-            return Unit.Value;
+
         }
     }
 }
