+++ source/dotnet-cqrs-petclinic-rest/PetClinic.Application/PetTypes/DeletePetType/DeletePetTypeCommandHandler.cs
@@ -3,10 +3,11 @@ using System.Threading;
 using System.Threading.Tasks;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
+using PetClinic.Domain.Common.Exceptions;
 using PetClinic.Domain.Repositories;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
-[assembly: IntentTemplate("Intent.Application.MediatR.CommandHandler", Version = "1.0")]
+[assembly: IntentTemplate("Intent.Application.MediatR.CommandHandler", Version = "2.0")]
 
 namespace PetClinic.Application.PetTypes.DeletePetType
 {
@@ -22,11 +23,16 @@ namespace PetClinic.Application.PetTypes.DeletePetType
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
-        public async Task<Unit> Handle(DeletePetTypeCommand request, CancellationToken cancellationToken)
+        public async Task Handle(DeletePetTypeCommand request, CancellationToken cancellationToken)
         {
             var existingPetType = await _petTypeRepository.FindByIdAsync(request.Id, cancellationToken);
+            if (existingPetType is null)
+            {
+                throw new NotFoundException($"Could not find PetType '{request.Id}'");
+            }
+
             _petTypeRepository.Remove(existingPetType);
-            return Unit.Value;
+
         }
     }
 }
