+++ source/dotnet-cqrs-petclinic-rest/PetClinic.Application/Visits/DeleteVisit/DeleteVisitCommandHandler.cs
@@ -3,10 +3,11 @@ using System.Threading;
 using System.Threading.Tasks;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
+using PetClinic.Domain.Common.Exceptions;
 using PetClinic.Domain.Repositories;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
-[assembly: IntentTemplate("Intent.Application.MediatR.CommandHandler", Version = "1.0")]
+[assembly: IntentTemplate("Intent.Application.MediatR.CommandHandler", Version = "2.0")]
 
 namespace PetClinic.Application.Visits.DeleteVisit
 {
@@ -22,11 +23,16 @@ namespace PetClinic.Application.Visits.DeleteVisit
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
-        public async Task<Unit> Handle(DeleteVisitCommand request, CancellationToken cancellationToken)
+        public async Task Handle(DeleteVisitCommand request, CancellationToken cancellationToken)
         {
             var existingVisit = await _visitRepository.FindByIdAsync(request.Id, cancellationToken);
+            if (existingVisit is null)
+            {
+                throw new NotFoundException($"Could not find Visit '{request.Id}'");
+            }
+
             _visitRepository.Remove(existingVisit);
-            return Unit.Value;
+
         }
     }
 }
